Moved the Google Maps (geolocation) logic out of minnesota.js and into geo.js. Also renamed MinnesotaTour.tourCoords to MinnesotaTour.cities.
Added some exception handling in the make_tour_request method so that it throws an exception if we have more cities than Google Maps will let us find a route for. At the moment this just throws up an alert window, but we might want to make that more sophisticated later.