Got rid of the redraw problem by not centering.
I used to set the center in myOptions, but that caused a redraw when
the tour was drawn, presumably because Google Maps centers on the tour,
overriding my centering. I'm a little worried that failure to control
the centering may cause the map to jitter during evolution, but we'll
deal with that if/when we get to it.